+++ src/main/java/net/silentchaos512/scalinghealth/utils/SHDifficulty.java
@@ -1,0 +1,14 @@
+package net.silentchaos512.scalinghealth.utils;
+
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
+
+/**
+ * Dummy class for compiling and legacy scaling health support
+ */
+public class SHDifficulty {
+
+    public static double areaDifficulty(World world, BlockPos pos) {
+        return 0;
+    }
+}